Move `DbPruner` tests to separate file
